Add custom configurable caching.
If the cache property is set to the boolean value 'true' then it defaults to
the old behavior of caching via memjs. If the cache property is an object with
get(key, cb(err, result)) and set(key, value) functions then that will be used
to cache requests. This should allow caching of requests in an arbitrary
system (e.g. Redis) as well as supporting custom cache TTLs and expirations.